updated deck.js folder
